Add files via upload
**'Horse Racing - Tipster Bets'**
UID - Unique ID. Not used in the actual system. The dataset is housed in an access database - the UID keeps all the bets in order. You may find a use for it!
ID - Each tipsters bets are kept in date and time order with the ID being incremented for each new bet, for each tipster.
Tipster - The name of the tipster. Each tipsters bets are in order, followed by the next tipsters bets
Date - The date of the race. Previous experiments showed, to us anyway, that the date was not important, that it was the ID number that showed the system that each tipsters bets are a linear list. The date is still used as it has never been shown to affect results in a negative way.
Track - The name of the track. Some tipsters favour some tracks more than others and this affects their profit margin.
Horse - The name of the horse.
Bet Type - Is the bet a 'Win' bet or an 'Each Way' bet.
Odds - The odds that the tipster presenting the bet say they got for the bet. When you actually place a bet you rarely get the predicted odds. Would the system be better served by lowering the odds by 10% to 20% which would be more realistic?
Result - Did the bet Win or Lose.
Tipster Active - Is the tipster active - true or false